package the project
